# spec/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wallet Tracker REST API",
    "version": "1.0.0",
    "description": "Contract for the explicit REST operations implemented by the current server: health, authentication, categories, transactions, synthetic batch import, schedules, summary, settings and AI drafts. Descriptions record implementation rules that JSON Schema cannot fully express. Document validation alone is not evidence of executed HTTP scenarios: check:openapi validates this contract and test:contract separately exercises disposable offline/demo servers, reporting observed and unexercised statuses. No full-project regression or GitHub CI result is implied by this document. OpenAPI 3.0.3 is selected for parser compatibility, not claimed to be the latest version. This describes a personal project, not a production service."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local application; PORT is configurable. No hosted demo is promised."
    }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "registerAccount",
        "tags": ["Authentication"],
        "summary": "Create an account and return a bearer token",
        "description": "Public endpoint. Email is trimmed, checked with the application's regex, then lowercased for storage and uniqueness. Password requires at least 8 JavaScript UTF-16 code units and at most 72 UTF-8 bytes, not 72 characters. Account and starter categories are created atomically. Existing normalized email returns 409, including a concurrent UNIQUE conflict. Extra JSON fields are ignored, not rejected. The token is signed with HS256 and expires after 30 days.",
        "security": [],
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/TokenCreated" },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/categories": {
      "get": {
        "operationId": "listCategories",
        "tags": ["Categories"],
        "summary": "List categories owned by the authenticated account",
        "description": "Returns all own categories ordered by id, including the current nullable monthly_limit_cents. No pagination or spending filter. Requires bearer authentication through the categories mount.",
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "responses": {
          "200": {
            "description": "Own category list, possibly empty.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Category" } } } }
          },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "operationId": "createCategory",
        "tags": ["Categories"],
        "summary": "Create an own category without a budget",
        "description": "Name is trimmed and must have 1-60 JavaScript UTF-16 code units after trim. The trimmed name is unique per account, case-sensitive. Icon omission/null stores null; every other JSON value is converted with String(value), not refused for having a non-string type. Extra POST fields are ignored, including a submitted monthly_limit_cents; the new category has no limit. This is not the strict PATCH shape policy.",
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CategoryCreateRequest" } } }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/CategoryCreated" },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/summary": {
      "get": {
        "operationId": "getMonthSummary",
        "tags": ["Summary"],
        "summary": "Read own monthly spending and budget states",
        "description": "Default is the server's current local month. Real expenses and paid instalments count; unpaid schedules and income do not add spending. Remaining balance is current monthly income minus that month's total, and may be negative. Overall and category limits are independent current values that also judge historical months. Server computes not_set for null, within for total<limit, reached for equality, exceeded for total>limit; limit_remaining_cents=limit-total may be negative. Limited categories appear even with zero spending; categories with neither spending nor a limit stay hidden. Uncategorised spending remains in totals with null category/name, null limit/remaining and not_set. Category rows are ordered by total descending then category id. Only the authenticated account's data is read.",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "pattern": "^\\d{4}-(0[1-9]|1[0-2])$" },
            "description": "Local YYYY-MM with month 01-12; omitted means server current month. No explicit year range beyond four digits. Unknown query parameters are ignored."
          }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/MonthSummary" },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/settings": {
      "get": {
        "operationId": "getSettings",
        "tags": ["Settings"],
        "summary": "Read own current income and overall spending budget",
        "description": "A fresh account, including an account with no settings row, reads monthly_income_cents=0 and monthly_limit_cents=null. These current settings apply to every queried month; no per-month history is stored.",
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/SettingsAccepted" },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "put": {
        "operationId": "putSettings",
        "tags": ["Settings"],
        "summary": "Save income and optionally set or clear the overall budget",
        "description": "monthly_income_cents is required, a number of integer cents from 0 through 100000000; no string conversion. monthly_limit_cents is optional: omission preserves the existing limit (or null on a new row); explicit null clears it; zero is a real zero budget. Limits use the same inclusive ceiling. Unknown PUT fields are ignored, not refused like PATCH. Category limits are not changed or checked against the overall limit. Returns the saved current settings.",
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SettingsWriteRequest" } } }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/SettingsAccepted" },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/transactions": {
      "get": {
        "operationId": "listTransactions",
        "tags": ["Transactions"],
        "summary": "List own expenses within an inclusive local-date range",
        "description": "Defaults from/to independently to current server-local month's boundaries; real calendar dates required and from<=to. Sorted spent_on descending, then id descending. limit defaults50, accepts0-200 after Number conversion; offset defaults0, accepts nonnegative integers with no explicit ceiling. Empty limit/offset use defaults. Optional category_id uses ownership validation. Empty category_id maps null, but SQL category_id=null matches no rows; it is not an uncategorised filter. Unknown query keys ignored.",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "name": "from", "in": "query", "schema": { "$ref": "#/components/schemas/LocalDate" }, "description": "Inclusive start, omitted=current month start." },
          { "name": "to", "in": "query", "schema": { "$ref": "#/components/schemas/LocalDate" }, "description": "Inclusive end, omitted=current month end." },
          { "name": "category_id", "in": "query", "schema": { "type": "string" }, "description": "Number-converted owned positive integer ID; foreign/missing404, invalid400. Empty string mapsnull and matches no rows." },
          { "name": "limit", "in": "query", "schema": { "type": "string", "default": "50" }, "description": "Number(value) must be integer0-200; empty string defaults50." },
          { "name": "offset", "in": "query", "schema": { "type": "string", "default": "0" }, "description": "Number(value) must be nonnegative integer; empty string defaults0. No explicit upper ceiling." }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/TransactionList" },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "operationId": "createTransaction",
        "tags": ["Transactions"],
        "summary": "Save an expense, optionally with an account-scoped operation key",
        "description": "Positive integer cents only. category_id uses Number conversion and own-category lookup; note uses String(non-null JSON). Omitted/null unkeyed spent_on defaults to server-local today. New expenses must use a real date no later than tomorrow; no explicit lower-year bound. Extra POST fields ignored. Budgets never block saving. With Idempotency-Key, spent_on must be explicit/non-null; canonical [amount,category-or-null,date,note-or-null] identifies the payload, independent of JSON key order and ignored extras. New keyed saves recheck ownership/date bound atomically. Same-key/same-payload replay returns original201 snapshot even after an edit, with Idempotency-Replayed=true; replay does not recheck current date bounds/category ownership. Changed payload409 IDEMPOTENCY_CONFLICT. Deleted original leaves a409 IDEMPOTENCY_REPLAY_UNAVAILABLE tombstone. Without a key, a repeated request can create duplicates. Keys are account-scoped and retained for the account lifetime.",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransactionCreateRequest" } } }
        },
        "responses": {
          "201": {
            "description": "New expense or original keyed snapshot. Header absent for a new save, present=true on replay.",
            "headers": {
              "X-Request-Id": { "$ref": "#/components/headers/RequestId" },
              "Idempotency-Replayed": { "$ref": "#/components/headers/IdempotencyReplayed" }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Transaction" } } }
          },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/KeyedConflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/schedules": {
      "get": {
        "operationId": "listSchedules",
        "tags": ["Schedules"],
        "summary": "List own subscriptions and loans",
        "description": "Only exact query active=1 selects active rows; every other value or omission returns both active and inactive rows without validation error. Sorted by computed next_due. next_due is based on the later of server-local today and starts_on; day31 clamps to the last existing day in a short month. Loan remaining counts/cents are computed, not stored; subscriptions returnnull for these fields. No schedule DELETE endpoint exists.",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "name": "active", "in": "query", "schema": { "type": "string" }, "description": "Exactly1 filters active only. All other values/omission return all own rows." }
        ],
        "responses": {
          "200": {
            "description": "Own schedules ordered by next_due, possibly empty.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Schedule" } } } }
          },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "operationId": "createSchedule",
        "tags": ["Schedules"],
        "summary": "Create a subscription or finite loan",
        "description": "Required trimmed name, positive integer amount_cents, day_of_month1-31 and real local starts_on. category_id omission/null/empty string means no category; otherwise Number-converted owned positive integer. total_count omission/null creates a subscription; integer1-1200 creates a loan. Amount times loan count must remain a safe integer. Unknown POST fields, including client counters/active, are ignored; initial paid_count=0,active=1. No past/future bound is imposed on starts_on beyond calendar validity.",
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScheduleCreateRequest" } } }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/ScheduleAccepted" },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "operationId": "loginAccount",
        "tags": ["Authentication"],
        "summary": "Sign in and return a bearer token",
        "description": "Public endpoint. Email is trimmed/lowercased for lookup. Missing, non-string or rejected credentials all answer 401 with the same message; login does not reapply registration's password length or email regex checks. Extra JSON fields are ignored. Before bcrypt, failed credentials are limited by email OR client IP: default 5 failures in a fixed 15-minute window, configurable locally. The next attempt after the limit returns 429 with Retry-After seconds. A successful login clears the tracked email and IP failures. Malformed JSON is rejected globally with 400, not counted as a failed credential. All 401 failures, including non-string credentials, are recorded.",
        "security": [],
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" }
        ],
        "requestBody": {
          "required": false,
          "description": "Credential fields may be omitted; missing or non-string credentials return 401 (or 429 if already blocked).",
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/TokenAccepted" },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "HS256 JWT returned by registration/login. The application's Authorization header uses case-sensitive Bearer followed by a token. Missing, invalid or expired token returns 401. The three public operations explicitly override this root requirement with security=[]."
      }
    },
    "schemas": {
...
      "ScheduleCreateRequest": {
        "type": "object",
        "additionalProperties": true,
        "required": ["name", "amount_cents", "day_of_month", "starts_on"],
        "properties": {
          "name": { "$ref": "#/components/schemas/TrimmedName" },
          "amount_cents": { "$ref": "#/components/schemas/AmountCents" },
          "category_id": { "$ref": "#/components/schemas/CategoryIdInput" },
          "day_of_month": { "$ref": "#/components/schemas/DayOfMonth" },
          "starts_on": { "$ref": "#/components/schemas/LocalDate" },
          "total_count": { "$ref": "#/components/schemas/TotalCount" }
        },
        "description": "POST ignores additional fields. Loan amount*count must be a safe integer; both bounds currently keep the product safe. No relative date bound on starts_on.",
        "example": { "name": "Synthetic subscription", "amount_cents": 999, "day_of_month": 31, "starts_on": "2026-09-28", "total_count": null }
      },
...
      "LocalDate": {
        "type": "string",
        "format": "date",
        "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
        "description": "Real local calendar date YYYY-MM-DD, no timezone conversion. Source validates month/day including leap years and imposes no additional year ceiling within four digits."
      },
...
      "Transaction": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "amount_cents", "category_id", "spent_on", "note", "schedule_id", "created_at"],
        "properties": {
          "id": { "type": "integer", "minimum": 1 },
          "amount_cents": { "$ref": "#/components/schemas/AmountCents" },
          "category_id": { "type": "integer", "nullable": true, "minimum": 1 },
          "spent_on": { "$ref": "#/components/schemas/LocalDate" },
          "note": { "type": "string", "nullable": true },
          "schedule_id": { "type": "integer", "nullable": true, "minimum": 1 },
          "created_at": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}:\\d{2}$",
            "description": "SQLite datetime('now') UTC timestamp with a space, not RFC3339/date-time. Do not parse stored local spent_on through timezone conversion."
          }
        }
      },
      "TransactionCreateRequest": {
        "type": "object",
        "additionalProperties": true,
        "required": ["amount_cents"],
        "properties": {
          "amount_cents": { "$ref": "#/components/schemas/AmountCents" },
          "category_id": { "$ref": "#/components/schemas/CategoryIdInput" },
          "spent_on": {
            "type": "string", "nullable": true, "format": "date", "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "description": "Unkeyed omission/null defaults to today. With Idempotency-Key, explicit non-null valid date required. New saves no later than server tomorrow; replay preserves original date. Header-dependent requirement and relative bound are prose rules, not supplied defaults."
          },
          "note": { "$ref": "#/components/schemas/StringifiedInput" }
        },
        "example": { "amount_cents": 1250, "category_id": "1", "spent_on": "2026-09-28", "note": "Synthetic lunch" }
      },
...
      "IncomeCents": {
        "type": "integer",
        "minimum": 0,
        "maximum": 100000000,
        "description": "Nonnegative monthly income in EUR cents; zero is allowed."
      },
...
      "Category": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "name", "icon", "monthly_limit_cents"],
        "properties": {
          "id": { "type": "integer", "minimum": 1 },
          "name": { "type": "string", "minLength": 1 },
          "icon": { "type": "string", "nullable": true },
          "monthly_limit_cents": { "$ref": "#/components/schemas/BudgetLimit" }
        },
        "example": { "id": 1, "name": "Synthetic category", "icon": null, "monthly_limit_cents": 5000 }
      },
      "CategoryCreateRequest": {
        "type": "object",
        "additionalProperties": true,
        "required": ["name"],
        "properties": {
          "name": { "$ref": "#/components/schemas/TrimmedName" },
          "icon": { "$ref": "#/components/schemas/StringifiedInput" }
        },
        "description": "POST ignores unknown fields; a posted monthly_limit_cents does not create a budget. Name and optional icon are the only values used."
      },
...
      "SettingsWriteRequest": {
        "type": "object",
        "additionalProperties": true,
        "required": ["monthly_income_cents"],
        "properties": {
          "monthly_income_cents": { "$ref": "#/components/schemas/IncomeCents" },
          "monthly_limit_cents": { "$ref": "#/components/schemas/BudgetLimit" }
        },
        "description": "PUT requires income and ignores extra fields. Omitted limit preserves the stored value, null removes it, zero sets a real limit. Numbers are not coerced from strings.",
        "example": { "monthly_income_cents": 150000, "monthly_limit_cents": null }
      },
...
      "RegisterRequest": {
        "type": "object",
        "additionalProperties": true,
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "description": "After trim(), must match /^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$/. Stored lowercased. No stricter email format validator or explicit length ceiling is applied. Schema/prose does not invent RFC email validation."
          },
          "password": {
            "type": "string",
            "description": "JavaScript string.length must be >=8 UTF-16 code units; Buffer.byteLength(password, 'utf8') must be <=72 bytes. JSON Schema character-length bounds do not express these rules, so they remain explicit prose and require executable boundary tests. Password is not trimmed."
          }
        }
      },
...
    },
    "parameters": {
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "schema": { "type": "string", "pattern": "^[A-Za-z0-9-]{8,64}$" },
        "description": "Only expense POST supports it. Account-scoped lifetime operation key. Explicit non-null spent_on required. Same canonical payload replays201, changed payload409, deleted original409. No key means no deduplication guarantee."
      },
      "RequestId": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "schema": { "type": "string" },
        "description": "A lowercase UUID matching the response header pattern is echoed. Any other value, including uppercase or non-UUID text, is replaced by a generated UUID; it does not cause 400."
      }
    },
    "headers": {
      "IdempotencyReplayed": {
        "description": "Present with value true only on a successful keyed expense replay; absent on a new save. Conditional presence must be tested by the runner, not treated as globally required.",
        "schema": { "type": "string", "enum": ["true"] }
      },
      "RequestId": {
        "description": "Always present. Caller lowercase UUID is echoed; otherwise a new UUID is generated. The x-required extension is for the later response validator, not an OpenAPI required keyword.",
        "x-required": true,
        "schema": {
          "type": "string",
          "format": "uuid",
          "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
        }
      },
...
    },
    "responses": {
...
      "TransactionList": {
        "description": "Own filtered expense page and total matching count.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransactionList" } } }
      },
...
      "KeyedConflict": {
        "description": "IDEMPOTENCY_CONFLICT for changed canonical payload or IDEMPOTENCY_REPLAY_UNAVAILABLE when the original expense was deleted.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KeyedError409" } } }
      },
      "CategoryCreated": {
        "description": "Own category created, monthly_limit_cents is null; no posted budget was applied.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Category" } } }
      },
...
      "MonthSummary": {
        "description": "Authenticated account's monthly totals and read-time overall/category budget states.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Summary" } } }
      },
      "SettingsAccepted": {
        "description": "Current own income and overall limit; no computed budget status field in this response.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Settings" } } }
      },
      "NotFound": {
        "description": "NOT_FOUND: no requested category, transaction or schedule belongs to the account. Missing and foreign owned IDs both return404, not403.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error404" } } }
      },
      "TokenCreated": {
        "description": "Account created; token returned.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Token" } } }
      },
...
      "InvalidRequest": {
        "description": "VALIDATION_FAILED: invalid values, date range, ID conversion or PATCH shape, or malformed JSON rejected before a handler. Login bad/missing credential fields instead use401. The global parser also applies before auth and GET handlers.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error400" } } }
      },
      "Unauthorized": {
        "description": "UNAUTHORIZED: login uses the same E-mail or password is incorrect message for absent/non-string/rejected credentials. Protected mounts refuse absent/malformed bearer headers or invalid/expired tokens with their authentication message. No account-existence disclosure.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error401" } } }
      },
      "Conflict": {
        "description": "CONFLICT: duplicate normalized account email or own case-sensitive category name; inactive schedule payment or total_count below already paid_count. Keyed expense conflicts use a separate status-specific schema.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error409" } } }
      },
      "PayloadTooLarge": {
        "description": "PAYLOAD_TOO_LARGE: the global Express JSON parser's default 100 KiB body limit is exceeded.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error413" } } }
      },
...
      "InternalError": {
        "description": "INTERNAL: Unexpected server error. Documented possible unhandled failure; not induced or observed by this document-validation slice.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error500" } } }
      }
    }
  }
}